Generative AI/2_speech_llm/pyproject.toml
[project]
name = "speech-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.118.0",
    "groq>=0.32.0",
    "gtts>=2.5.4",
    "numpy>=2.3.3",
    "openai-whisper>=20250625",
    "pyaudio>=0.2.14",
    "pyttsx3>=2.99",
    "rich>=14.1.0",
    "soundfile>=0.13.1",
    "speechrecognition>=3.14.3",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
    "webrtcvad>=2.0.10",
]
